requirements.txt
web3>=6.0.0
PyYAML>=6.0
colorama>=0.4.4